confluence-function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-functions-durable
azure-storage-blob
azure-identity
azure-keyvault-secrets
azure.data.tables
azure.mgmt.resource
atlassian
pandas
pyyaml